Create ScoreSender Provider
