+++ smartPhone/src/app/app.module.ts
@@ -47,6 +47,7 @@ import { DatabaseService } from '../service/database-service';
 import { MedicalPage } from '../pages/medical/medical';
 import { TestResultsPage } from '../pages/test-results/test-results';
 import { ScoreProvider } from '../providers/score/score';
+import { ScoreSender } from '../providers/score-sender/score-sender';
 
 @NgModule({
   declarations: [
@@ -127,7 +128,8 @@ import { ScoreProvider } from '../providers/score/score';
     SQLite,
     DatabaseService,
     ScoreProvider,
-    {provide: ErrorHandler, useClass: IonicErrorHandler}
+    {provide: ErrorHandler, useClass: IonicErrorHandler},
+    ScoreSender
   ]
 })
