+++ smartPhone/src/data/constants.ts
@@ -1,4 +1,5 @@
 
 
 export const API_ENDPOINT= 'http://ec2-3-17-143-36.us-east-2.compute.amazonaws.com:5000';
-export const UPDATE_TABLE_URL = API_ENDPOINT+'/api/integracion/table/update';
+export const INSERT_REGISTRY_URL = API_ENDPOINT+'/api/integracion/table/insert';
+export const UPDATE_REGISTRY_URL = API_ENDPOINT+'/api/integracion/table/update';
